testapp/requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
certifi==2019.3.9
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
dj-database-url==0.4.2
Django==2.1
django-cors-headers==2.2.0
django-filter==2.0.0
django-webpack-loader==0.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ecdsa==0.13
future==0.16.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.7.3.2
pycodestyle==2.4.0
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.6.4
python-jose==1.4.0
python-magic==0.4.15
pytz==2018.5
requests==2.22.0
six==1.11.0
sqlparse==0.3.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==3.3.1
wrapt==1.10.11
